ciscolib compiles

+++ include/ciscolib-const.h
@@ -3,37 +3,37 @@
 * (c)2021 pocketlinux32, Under GPLv3 *
 * Ciscolib Constants Header File     *
 \************************************/
-#define CISCO_INT_F0 uint8_t(0)
-#define CISCO_INT_G0 uint8_t(1)
-#define CISCO_INT_G00 uint8_t(2)
-#define CISCO_INT_G01 uint8_t(3)
-#define CISCO_INT_S00 uint8_t(4)
-#define CISCO_INT_S01 uint8_t(5)
-#define CISCO_INT_VLAN uint8_t(6)
-#define CISCO_INT_PORTCH uint8_t(7)
+#define CISCO_INT_F0 0
+#define CISCO_INT_G0 1
+#define CISCO_INT_G00 2
+#define CISCO_INT_G01 3
+#define CISCO_INT_S00 4
+#define CISCO_INT_S01 5
+#define CISCO_INT_VLAN 6
+#define CISCO_INT_PORTCH 7
 
-#define CISCO_MODE_ACCESS uint8_t(16)
-#define CISCO_MODE_TRUNK uint8_t(17)
-#define CISCO_MODE_ACTIVE uint8_t(18)
-#define CISCO_MODE_PASSIVE uint8_t(19)
-#define CISCO_MODE_DESIRABLE uint8_t(20)
-#define CISCO_MODE_AUTO uint8_t(21)
-#define CISCO_MODE_ON uint8_t(22)
-#define CISCO_MODE_IN_PORTCH uint8_t(23)
+#define CISCO_MODE_ACCESS 16
+#define CISCO_MODE_TRUNK 17
+#define CISCO_MODE_ACTIVE 18
+#define CISCO_MODE_PASSIVE 19
+#define CISCO_MODE_DESIRABLE 20
+#define CISCO_MODE_AUTO 21
+#define CISCO_MODE_ON 22
+#define CISCO_MODE_IN_PORTCH 23
 
-#define CISCO_MODTYPE_TYPE uint8_t(32)
-#define CISCO_MODTYPE_MODE uint8_t(33)
-#define CISCO_MODTYPE_PORTS uint8_t(34)
-#define CISCO_MODTYPE_DESC uint8_t(35)
-#define CISCO_MODTYPE_ALLOW_VLAN uint8_t(36)
-#define CISCO_MODTYPE_IP_ADDR uint8_t(37)
-#define CISCO_MODTYPE_SUBMASK uint8_t(38)
-#define CISCO_MODTYPE_GATEWAY uint8_t(39)
+#define CISCO_MODTYPE_TYPE 32
+#define CISCO_MODTYPE_MODE 33
+#define CISCO_MODTYPE_PORTS 34
+#define CISCO_MODTYPE_DESC 35
+#define CISCO_MODTYPE_ALLOW_VLAN 36
+#define CISCO_MODTYPE_IP_ADDR 37
+#define CISCO_MODTYPE_SUBMASK 38
+#define CISCO_MODTYPE_GATEWAY 39
 
-#define CISCO_ERROR_INVALID_ACTION uint8_t(48)
-#define CISCO_ERROR_INVALID_VALUE uint8_t(49)
-#define CISCO_ERROR_BUFFER_OVERFLOW uint8_t(50)
-#define CISCO_ERROR_OUT_OF_RANGE uint8_t(51)
-#define CISCO_ERROR_PL32LIB_GC uint8_t(52)
-#define CISCO_ERROR_MISMATCHED_IPVER uint8_t(53)
+#define CISCO_ERROR_INVALID_ACTION 48
+#define CISCO_ERROR_INVALID_VALUE 49
+#define CISCO_ERROR_BUFFER_OVERFLOW 50
+#define CISCO_ERROR_OUT_OF_RANGE 51
+#define CISCO_ERROR_PL32LIB_GC 52
+#define CISCO_ERROR_MISMATCHED_IPVER 53
